+++ sprite_system.h
@@ -19,7 +19,7 @@ void sprite_free(int id);
 
 int sprite_width(int id);
 int sprite_height(int id);
-FColor sprite_fsample(int id, double x, double y);
+FColor sprite_fsample(int id, float x, float y);
 FColor sprite_sample(int id, int x, int y);
 
 FColor fcolor_from(unsigned char c[4]);
